Moved ParticleLoopImplementation over to runtime library where possible.
